LPS-185207 It is an utility class. In this way it can be used by test modules too. No need to move it to portal-kernel since it is only for control tables so it does not impact modules.
