+++ portal-impl/src/com/liferay/portal/upgrade/util/UpgradePartitionedControlTable.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.upgrade.v7_4_x;
+package com.liferay.portal.upgrade.util;
 
 import com.liferay.portal.db.partition.DBPartitionUtil;
 import com.liferay.portal.kernel.db.partition.DBPartition;
